+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/General/Legend.cs
@@ -25,6 +25,27 @@ public class Legend : ICategory
         }
     }
 
+    public class LegendWithRTLText : RecipeBase
+    {
+        public override string Name => "Right To Left (RTL) text support";
+        public override string Description => "Enabling Right To Left (RTL) support " +
+            "correctly renders text in RTL languages.";
+
+        [Test]
+        public override void Execute()
+        {
+            LabelStyle.RTLSupport = true; // enable right-to-left text support
+
+            var sig1 = myPlot.Add.Signal(Generate.Sin(51));
+            sig1.LegendText = "אמת"; // example right-to-left text
+
+            var sig2 = myPlot.Add.Signal(Generate.Cos(51));
+            sig2.LegendText = "English"; // example left-to-right text
+
+            myPlot.ShowLegend();
+        }
+    }
+
     public class ManualLegend : RecipeBase
     {
         public override string Name => "Manual Legend Items";