+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WPF Demo/DemoWindows/Quickstart.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Windows;
+﻿using ScottPlot;
+using System.Windows;
 
 namespace WPF_Demo.DemoWindows;